app/requirements.txt
huey==2.5.0
redis==5.0.1
Django==5.0.1
pillow==10.2.0
python-dotenv==1.0.0
psycopg2-binary==2.9.9
django-money==0.11.4
django-cors-headers==4.3.1
django-ninja==1.1.0
django-ninja-jwt==5.2.10